requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
scipy==1.11.3
Pillow==10.0.1
albumentations==1.3.1
timm==0.9.7
python-multipart==0.0.6
tqdm==4.66.1
pyyaml==6.0.1
supabase==2.3.4
python-dotenv==1.0.0
